NP-1668: move stacktrace method to ExceptionUtils
