+++ src/main/java/nva/commons/handlers/RestRequestHandler.java
@@ -1,7 +1,7 @@
 package nva.commons.handlers;
 
 import static java.util.Objects.isNull;
-import static nva.commons.utils.StringUtils.stackTraceInSingleLine;
+import static nva.commons.utils.ExceptionUtils.stackTraceInSingleLine;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestStreamHandler;
