+++ src/main/java/nva/commons/utils/StringUtils.java
@@ -2,10 +2,6 @@ package nva.commons.utils;
 
 import static java.util.Objects.isNull;
 
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.util.stream.Stream;
-
 public final class StringUtils {
 
     public static final String DOUBLE_WHITESPACE = "\\s\\s";
@@ -54,21 +50,5 @@ public final class StringUtils {
      */
     public static String replaceWhiteSpacesWithSpace(String str) {
         return str.replaceAll("\\s", " ");
-    }
-
-    /**
-     * Returns the stacktrace in one line. It replaces all whitespaces with space and removes multiple whitespaces.
-     *
-     * @param e the Exception
-     * @return the Stacktrace String.
-     */
-    public static String stackTraceInSingleLine(Exception e) {
-        StringWriter sw = new StringWriter();
-        e.printStackTrace(new PrintWriter(sw));
-        String exceptionString = sw.toString();
-        return Stream.of(exceptionString)
-            .map(StringUtils::removeMultipleWhiteSpaces)
-            .map(StringUtils::replaceWhiteSpacesWithSpace)
-            .collect(SingletonCollector.collect());
     }
 }
